+++ src/components/LeaderboardMember.astro
@@ -9,7 +9,7 @@ interface Props {
 const { member } = Astro.props;
 
 import { getDollarsPerDev, fmtCurrency } from '../memberFormatting.ts';
-import type { MemberWithId } from "../content/config.ts";
+import type { MemberWithId } from "../schemas/members.ts";
 ---
 
 <tr>
