+++ src/content/articles/join-the-pledge.md
@@ -1,0 +1,101 @@
+---
+title: 'Join the Pledge'
+excerpt: "Today we officially launch the Open Source Pledge."
+author: 'David Cramer'
+authorImageSrc: '/images/authors/cramer.webp'
+isDraft: false
+publishDate: '2024-10-08'
+---
+
+<div class="highlight-box">
+  This post originally appeared
+  <a
+    href="https://blog.sentry.io/join-the-pledge/"
+  >on blog.sentry.io</a>.
+</div>
+
+Today we officially launch the Open Source Pledge.
+
+The Pledge started as an idea some years back: what if we could give back to Open Source on behalf of every employee at
+Sentry? We threw around a number of ideas on how we might do that, but none of them seemed like they’d achieve the level
+of impact we wanted. We always had two core goals with it:
+
+1. Pay maintainers, directly.
+2. Do it sustainably, and scale it with our growth.
+
+My earliest thought in this space centered around a form of donation matching. The hope was we could take something like
+GitHub Sponsors and match employees' contributions to Open Source maintainers. That posed a number of challenges, but
+the biggest risk was participation. Not everyone cares about Open Source (that’s fine!), so donation matching breaks
+down with reduced participation. Instead, we decided to do direct funding, driven by a variety of inputs (our dependency
+graph, projects people voted on, and guidance from engineering leadership).
+
+That program is what you’ve [seen from us publicly][sentry-500], running it for three consecutive years. Each year we
+increased the funding amount based on Sentry’s own financial growth. It became such a no-brainer within Sentry’s
+leadership that we have aggressively increased the funding each year, even beyond our original targets. With the success
+of that, we set off to take that program, codify it, and bring it to other companies to see if we could turn this into a
+bigger thing.
+
+## Lead by Example
+
+As we started talking about this, thinking about how we might turn it into something bigger with more impact, I was
+reminded of a number of scenarios that I had personally experienced when speaking with other founders. I regularly speak
+at a variety of events, many where Open Source is a key part of the narrative. They’re generally filled with
+venture-backed founders telling their story of why Open Source matters to their business, why they believe, and invest
+in it. One thing that was commonly expressed by these folks was the sustainability challenges in the industry. “Great!”
+I thought, “We have a solution!”
+
+Not a single one of those founders did anything more than talk about the problem. Sure, maybe they throw a few bucks at
+one of the big foundations, and they almost certainly fund investments into their own ecosystem (their commercial
+interests). Unfortunately, they rarely do anything measurable that improves the thing they claim to care so deeply
+about. Worse, some of these people (often from the largest tech companies), have a laundry list of excuses for why
+giving money to people is hard.
+
+So we decided to try and do something about it. That something is the Open Source Pledge. We don't think it’s the only
+solution, nor do we think it’s the only way to give back, but we do believe giving cash money to maintainers is an
+appropriate way to show your thanks, to recognize their hard work, the value they create for you. Maybe, just maybe,
+we’ll do our small part in encouraging the maintainers to keep putting up with us in the enormous ecosystem we rely on.
+
+## How to Pledge
+
+If you’ve made it this far, there are probably two questions you have.
+
+**First, what does it take to join the Pledge?** You commit (with receipts) to giving 2,000 USD to your dependencies,
+per engineer you employ, every single year.
+
+Sentry has ~135 engineers on staff, meaning our minimum commitment is $270,000 this year. Think about that in the
+context that Sentry generates more than $100mm in recurring revenue. It’s a fraction of what we spend in any given
+calendar month on digital advertising. It's a pretty modest amount in the grand scheme of things, but it's enough to
+have genuine impact.
+
+**Second, what’s the ROI your company gets from joining?** It’s marketing. It’s your brand. It's top-of-funnel. It’s
+software security. It’s free software. It’s Open Source.
+
+We see ROI in two major ways: brand marketing, and the supply chain. You may care about one more than the other, so
+choose whichever helps you get over the finish line.
+
+From the supply chain angle, you’re encouraging maintainers of the software you use to continue to provide support.
+You’re telling them that you value their work, and the contribution is there to encourage them to keep contributing to
+it. At the very least, you’re giving them a big thank you, setting the tone for future generations of maintainers.
+You’re thanking them for free software. Both of those improve the efficiency of your R&D investments.
+
+From the brand angle, it's what you make of it. If it's a space you care about, you are putting your money where your
+mouth is, and your audience will recognize that. You buy products from brands that you connect with, and if your
+customers care about Open Source, you’re giving them one more reason to care about you over your competition. You’re
+also putting yourself out there, attracting new eyes to your brand that may not have heard about you before.
+
+We don’t yet know if the Pledge will be successful, but I’m thrilled with the number of people who have decided to
+support the program ([directly with funds][members], and [indirectly with broadcasting and recruiting][endorsements]). I
+especially want to thank the people who have put in the hours to really get this off the ground: [Chad Whitacre][chad]
+and [Michael Selvidge][michael] from Sentry, as well as [Vlad-Stefan Harbuz][vlad] and [Ethan Arrowood][ethan].
+
+While Sentry is funding getting the program off the ground, we’re hoping it lasts well beyond us and turns into
+something much more. The industry is in a rocky place these days, and a little bit of effort from the people who can
+afford it can go a long way.
+
+[chad]: https://x.com/chadwhitacre_
+[endorsements]: https://opensourcepledge.com/endorsements/
+[ethan]: https://github.com/Ethan-Arrowood
+[members]: https://opensourcepledge.com/members/
+[michael]: https://www.linkedin.com/in/michaelselvidge/
+[sentry-500]: https://blog.sentry.io/we-just-gave-500-000-dollars-to-open-source-maintainers/
+[vlad]: https://vlad.website
