+++ src/layouts/Layout.astro
@@ -54,6 +54,9 @@ const { title, navless } = Astro.props;
           <li>
             <a class="sneaky" href="/members">Members</a>
           </li>
+          <li>
+            <a class="sneaky" href="/articles">Blog</a>
+          </li>
         </ul>
       </nav>
     ) }
@@ -63,6 +66,7 @@ const { title, navless } = Astro.props;
         <span><a class="sneaky" href="/join">Join</a></span>
         <span><a class="sneaky" href="/about">About</a></span>
         <span><a class="sneaky" href="/members">Members</a></span>
+        <span><a class="sneaky" href="/articles">Blog</a></span>
         <span><a class="sneaky" href="/endorsements">Endorsements</a></span>
         <span><a class="sneaky" href="/press">Press</a></span>
       </div>
@@ -421,6 +425,27 @@ const { title, navless } = Astro.props;
     padding-left: 0;
     li {
       list-style-position: inside;
+    }
+  }
+  .button {
+    display: inline-block;
+    margin: 1rem 0;
+    padding: 0.8rem 2rem;
+    color: var(--color-bg);
+    background-color: var(--color-primary);
+    border-radius: 0.5rem;
+    text-align: center;
+    text-decoration: none;
+    text-transform: uppercase;
+    font-weight: bold;
+    font-size: 1.2em;
+    &:hover {
+      background: var(--color-secondary);
+      color: var(--color-bg);
+    }
+    &:active {
+      background: var(--color-tertiary);
+      color: var(--color-bg);
     }
   }
   section {
