+++ src/pages/articles/index.astro
@@ -1,0 +1,43 @@
+---
+// © 2024 Vlad-Stefan Harbuz <vlad@vlad.website>
+// SPDX-License-Identifier: Apache-2.0
+
+import ArticleList from "../../components/ArticleList.astro";
+import Blob from "../../components/Blob.astro";
+import Button from "../../components/Button.astro";
+import Layout from "../../layouts/Layout.astro";
+---
+
+<Layout title="Blog Articles">
+  <div class="edge-blobs">
+    <Blob kind="grad-dots-01" top="8%" right="-2rem"></Blob>
+    <Blob kind="solid-fill-03" top="65%" left="-10rem"></Blob>
+  </div>
+  <div class="container">
+    <Blob kind="solid-fill-04" top="-15%" left="-16rem"></Blob>
+    <Blob kind="solid-fill-02" top="2%" right="-22rem"></Blob>
+    <Blob kind="grad-dots-02" top="30%" left="-20rem"></Blob>
+    <Blob kind="grad-dots-03" top="60%" right="-20rem"></Blob>
+
+    <main id="main-content">
+      <section>
+        <div class="text-center">
+          <h1>Articles</h1>
+        </div>
+        <ArticleList />
+      </section>
+
+      <section class="flex-center bottom-cta">
+        <div>
+          <div class="text-center">
+            <h2>Convinced?</h2>
+          </div>
+
+          <div class="text-center">
+            <Button href="/join">Join the Pledge</Button>
+          </div>
+        </div>
+      </section>
+    </main>
+  </div>
+</Layout>
