+++ src/pages/members/[id].astro
@@ -12,7 +12,7 @@ import {
 } from '../../memberFormatting.ts';
 import type { Map } from '../../util.ts';
 import memberRoles from "../../memberRoles.json"
-import type { Member } from "../../content/config.ts";
+import type { Member } from "../../schemas/members.ts";
 
 interface Props {
   member: Member;
